requirements.txt
certifi==2023.11.17
elastic-transport==8.12.0
elasticsearch==8.12.0
kafka-python==2.0.2
python-dotenv==1.0.1
urllib3==2.1.0
